## Change the implementation of the the services and controllers

+++ src/main/java/org/kemet/blogex/repository/BlogRepository.java
@@ -1,10 +1,10 @@
 package org.kemet.blogex.repository;
 
-import org.kemet.blogex.entity.Blog;
+import org.kemet.blogex.domain.Blog;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlogRepository extends JpaRepository<Blog, Integer> {
+public interface BlogRepository extends JpaRepository<Blog, Long> {
 
 }
