+++ src/main/java/org/kemet/blogex/repository/CommentaireRepository.java
@@ -1,10 +1,10 @@
 package org.kemet.blogex.repository;
 
-import org.kemet.blogex.entity.Commentaire;
+import org.kemet.blogex.domain.Commentaire;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentaireRepository extends JpaRepository<Commentaire, Integer> {
+public interface CommentaireRepository extends JpaRepository<Commentaire, Long> {
 	
 }
